netplan: avoid generating duplicate IPs
As of about Ubuntu 20.04 LTS, the netplan configuration defaults:

dhcp-identifier: duid

which generates a DHCP identifier from (basically) a hash of /etc/machine-id.
So if cloned disk images have the same /etc/machine-id, they will get
the same DHCP IP, and chaos will ensue (ie, the network connections will
be super unreliable, due to ARP bouncing the IP between multiple VMs).

Work around this in two ways:

1.  Set "dhcp-identifier: mac" (so it works like traditional DHCP clients); and
2.  Truncate /etc/machine-id so it gets recreated on first boot

See:

https://netplan.readthedocs.io/en/latest/netplan-yaml/

https://systemd.io/BUILDING_IMAGES/
https://wiki.debian.org/MachineId

Note that at least up to Ubuntu 22.04 LTS it is necesary to *truncate*
(not remove) /etc/machine-id, as the recovery steps run early on when
/etc is read only, so it needs to be able to mount a temporary replacement
over it and then commit it once the file system is read/write.  But
/var/lib/dbus/machine-id needs to be removed, as it is recreated as
a symlink to /etc/machine-id in that case.

(Only tested on Ubuntu 22.04 LTS, but seems like it should still work on
Ubuntu 20.04 LTS and mabye Ubuntu 18.04 LTS.)